requirements.txt
firebase-admin==5.2.0
flask-cors==3.0.10
google-cloud-firestore==2.3.4
google-cloud-logging==2.6.0
google-cloud-pubsub==2.9.0
google-cloud-secret-manager==2.8.0
numpy>=1.21.5
python_dateutil==2.8.2
pytz==2021.3
rdp==0.8
timezonefinder[numba]==5.2.0
